requirements.txt
# Web Framework
streamlit==1.33.0

# Core Scientific Stack
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2

# NLP Libraries
spacy==3.7.2
nltk==3.8.1
bertopic==0.16.0

# BERTopic Dependencies
hdbscan==0.8.33
umap-learn==0.5.5
torch==2.1.0
sentence-transformers==2.2.2
transformers==4.36.2
huggingface-hub==0.20.3

# Other Librariespandas==2.0.3
requests
newspaper3k
lxml[html_clean]

# Required spaCy Model
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl